requirements.txt
langchain
ipykernel
python-dotenv
langchain-openai
langchain-community
phidata
duckduckgo-search
streamlit
google-generativeai
faiss-cpu
arxiv
wikipedia
pypdf
langchain-groq
langgraph
langgraph-cli[inmem]